[Feat]: CustomerNPC -> CustomerBase/NormalCustomer/BadguyCustomer/GourmetCustomer/MichelinCustomer 로 Refactoring
