+++ Assets/Prefabs/NPCs/NPC_Bad.prefab
@@ -908,7 +908,7 @@ GameObject:
   m_Component:
   - component: {fileID: 4925318079550245093}
   - component: {fileID: 1649060776012356689}
-  - component: {fileID: 4152616506287534462}
+  - component: {fileID: 1069141251953515542}
   - component: {fileID: -8119461144268489840}
   m_Layer: 0
   m_Name: NPC_Bad
@@ -925,7 +925,7 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 5697596817086517855}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 1, y: 0.6, z: -7.5}
+  m_LocalPosition: {x: 0, y: 0.6, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children:
@@ -957,7 +957,7 @@ NavMeshAgent:
   m_BaseOffset: 0.5
   m_WalkableMask: 4294967295
   m_ObstacleAvoidanceType: 4
---- !u!114 &4152616506287534462
+--- !u!114 &1069141251953515542
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -966,27 +966,33 @@ MonoBehaviour:
   m_GameObject: {fileID: 5697596817086517855}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: a68d15752e8094451b62da36c3c5d087, type: 3}
+  m_Script: {fileID: 11500000, guid: b16da658e946347ac9718764258b5061, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  customerType: 0
+  customerType: 2
+  requiredDish:
+    <name>k__BackingField: 
+    <food>k__BackingField: 0
+    <tag>k__BackingField: 0
+    <id>k__BackingField: 0
+    <price>k__BackingField: 0
+    <isUnlocked>k__BackingField: 0
+    <isBought>k__BackingField: 0
+    <level>k__BackingField: 0
+    <prefab>k__BackingField: {fileID: 0}
+    <icon>k__BackingField: {fileID: 0}
   customerAgent: {fileID: 1649060776012356689}
   spawnPosition: {x: 1, y: 0.6, z: -7.5}
-  isFoodReceived: 0
-  patience: 90
-  patienceGauge: {fileID: 0}
-  orangeButton: {fileID: 21300000, guid: 95332da0e346cb345ac3ece62a649e61, type: 3}
-  orderButton: {fileID: 0}
-  menuText: {fileID: 0}
+  patienceGauge: {fileID: 1134768348397056033}
+  orderButton: {fileID: 5945055264131334109}
+  menuText: {fileID: 4856766188787070713}
   audioSource: {fileID: -8119461144268489840}
   eatSound: {fileID: 8300000, guid: c0431280aa8fc4384bdd2a0f5b93da8c, type: 3}
   drinkSound: {fileID: 8300000, guid: 58b24b4c0cf564f6c8e9b32f3781f1b1, type: 3}
-  HappyIcon: {fileID: 2800000, guid: 8c87a62bf109a4ac1984a4370900bc36, type: 3}
-  DisappointIcon: {fileID: 2800000, guid: 083263b96647d451f8f9a0a256e9c89b, type: 3}
-  BadguyIcon: {fileID: 2800000, guid: 7d43eb759852641c7830c6d17bbac1d1, type: 3}
-  DeadIcon: {fileID: 2800000, guid: 88ef467957313415dbd2518100ae4b7b, type: 3}
+  SuccessIcon: {fileID: 2800000, guid: 88ef467957313415dbd2518100ae4b7b, type: 3}
+  FailIcon: {fileID: 2800000, guid: 7d43eb759852641c7830c6d17bbac1d1, type: 3}
   foodDatabase: {fileID: 11400000, guid: 6a6a9e90f2e954b66b77ae48e1b2efe3, type: 2}
-  customerAnimator: {fileID: 0}
+  customerAnimator: {fileID: 7188421097076193914}
 --- !u!82 &-8119461144268489840
 AudioSource:
   m_ObjectHideFlags: 0
