+++ Assets/Scripts/CustomerBase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cb21b89870ea04bbd80af84a2994fd7c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
