+++ Assets/Scripts/GourmetCustomer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d21762d6fdbec4615a914ebad3993607
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
