+++ Assets/Scripts/MichelinCustomer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6b11e1157b507468db8a6c721dfe3d2f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
