+++ Assets/Scripts/NormalCustomer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 768839e49c00d4e118386de03f0a9987
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
